后台管理员搜索功能

+++ app/Http/Controllers/Api/UserController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Model\Agent;
+use App\Http\Model\Employee;
 use App\User;
+use Carbon\Carbon;
+use function foo\func;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -30,7 +34,7 @@ class UserController extends Controller
      */
     public function logoutApi()
     {
-        if (Auth::check()) {
+        if (Auth::check() && !config('app.debug')) {
             Auth::user()->AauthAcessToken()->delete();
         }
     }
@@ -60,5 +64,39 @@ class UserController extends Controller
         } catch (\Exception $e) {
             return self::jsonReturn([], 0, $e->getMessage());
         }
+    }
+
+
+    public function list(Request $request)
+    {
+
+        $keyword = $request->keyword;
+        $page_size = $request->input('size', self::PAGE_SIZE);
+
+
+        $query = User::orderByDesc('created_at')->with('agent')->with('role');
+        $range = $request->range;
+        $range = array_filter($range, function ($k, $value) {
+            return strlen($value) > 10;
+        }, ARRAY_FILTER_USE_BOTH);
+        if (count($range) == 2) {
+            $from_time = Carbon::parse($range[0]);
+            $to_time = Carbon::parse($range[1]);
+            $query->whereBetween('created_at', [$from_time, $to_time]);
+        }
+        if ($keyword) {
+            $likeString = "%$keyword%";
+            $agent_dis = Agent::where('agent_name', 'like', $likeString)
+                ->pluck('id')->all();
+            $meployee_ids = Employee::where('employee_name', 'like', $likeString)
+                ->pluck('id')->all();
+            $query->whereIn('agent_id', array_values($agent_dis))
+                ->orWhereIn('employee_id', $meployee_ids)
+                ->orWhere('phone', 'like', $likeString)
+                ->orWhere('real_name', 'like', $likeString);
+            $_GET['page'] = 1;
+        }
+        $list = $query->paginate($page_size);
+        return self::jsonReturn($list);
     }
 }
